fixed test for related count.

+++ backend/pkg/web/handler/resource_related_test.go
@@ -111,7 +111,7 @@ func (suite *ResourceRelatedHandlerTestSuite) TestResourceRelatedHandlerTestSuit
 
 	//assert
 	assert.EqualValues(suite.T(), http.StatusOK, w.Code)
-	assert.Equal(suite.T(), summary["TotalResources"], 3)
+	assert.Equal(suite.T(), summary["TotalResources"], float64(3))
 
 }
